djangoapp/requirements.txt
Django>=4.2.1,<4.3
Pillow
python-decouple==3.8
psycopg2-binary>=2.9.6,<2.10
gunicorn==20.1.0,<20.2
django-extensions==3.2.1,<3.3
six==1.16.0,<1.17
redis==4.5.4,<4.6
django-redis==5.2.0,<5.3
django-widget-tweaks==1.4.12,<1.5
django-pwa==1.1.0,<1.2
django-autoslug==1.9.9,<2.0
boto3==1.28.15,<2.29
channels==4.0.0,<4.1
daphne==4.0.0,<4.1
channels_redis==4.1.0,<4.2
Werkzeug==3.0.0,<3.1
pyOpenSSL==23.2.0,<23.3
django-storages==1.14.1,<1.15
# django-summernote==0.8.20.0,<0.9
celery==5.2.7
django-celery-beat==2.5.0
openai==1.30.1,<1.31
gevent==23.9.1,<24.0
python-telegram-bot
watchdog==4.0.0,<4.1
matplotlib==3.8.4,<3.9
djangorestframework==3.15.1,<3.16
requests==2.32.2,<2.33
requests[security]
google-api-python-client==2.145.0,<2.146
google-auth-httplib2==0.2.0,<0.3
google-auth-oauthlib==1.2.1,<1.3
